backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
python-multipart
pydantic
gunicorn
uvicorn
